# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Myme API",
    "version": "0.1.0",
    "description": "Typed data layer for structured personal data"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key or OAuth Token",
        "description": "Pass an API key (myme_k1_...) or OAuth access token (myme_at_...)"
      }
    },
...
  },
  "paths": {
...
    "/keys": {
      "post": {
        "tags": ["Keys"],
        "summary": "Create a new API key",
        "description": "Creates a new API key. In bootstrap mode (zero keys exist), no auth is required and the key is always admin.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1
                  },
                  "source": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "role": {
                    "type": "string",
                    "enum": ["admin", "member"]
                  },
                  "default_origin": {
                    "type": "string",
                    "enum": ["user", "ai", "worker"]
                  },
                  "default_library": {
                    "type": "boolean"
                  },
                  "type_permissions": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "enum": ["read", "write", "none"]
                    }
                  }
                },
                "required": ["label", "source"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "label": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "enum": ["admin", "member"]
                    },
                    "default_origin": {
                      "type": "string",
                      "enum": ["user", "ai", "worker"]
                    },
                    "default_library": {
                      "type": "boolean"
                    },
                    "type_permissions": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string",
                        "enum": ["read", "write", "none"]
                      }
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "last_used_at": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "key",
                    "label",
                    "source",
                    "role",
                    "default_origin",
                    "default_library",
                    "type_permissions",
                    "created_at",
                    "last_used_at"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Keys"],
        "summary": "List all API keys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "keys": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string"
                          },
                          "default_origin": {
                            "type": "string",
                            "enum": ["user", "ai", "worker"]
                          },
                          "default_library": {
                            "type": "boolean"
                          },
                          "type_permissions": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "last_used_at": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "label",
                          "source",
                          "role",
                          "default_origin",
                          "default_library",
                          "type_permissions",
                          "created_at",
                          "last_used_at"
                        ]
                      }
                    }
                  },
                  "required": ["keys"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
    "/keys/{id}": {
      "delete": {
        "tags": ["Keys"],
        "summary": "Revoke an API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Key revoked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [true]
                    }
                  },
                  "required": ["ok"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    },
...
  }
}